Finished Library class implementation and started the LibraryHandler class
Finished the implementation for the Library class, but it might undergo changes in the future. Also started the LibraryHandler class that is responsible for the correct execution of the whole program. As I have some questions about the requirements this is still a prototype class, but this should be the overall idea of its function.